np.int is deprecated

+++ model_angelo/utils/residue_constants.py
@@ -965,10 +965,10 @@ def _make_rigid_transformation_4x4(ex, ey, translation):
 # and an array with (restype, atomtype, coord) for the atom positions
 # and compute affine transformation matrices (4,4) from one rigid group to the
 # previous group
-restype_atomf_to_rigid_group = np.zeros([canonical_num_residues, atom_type_num], dtype=np.int)
+restype_atomf_to_rigid_group = np.zeros([canonical_num_residues, atom_type_num], dtype=np.int32)
 restype_atomf_mask = np.zeros([canonical_num_residues, atom_type_num], dtype=np.float32)
 restype_atomf_rigid_group_positions = np.zeros([canonical_num_residues, atom_type_num, 3], dtype=np.float32)
-restype_atomc_to_rigid_group = np.zeros([canonical_num_residues, num_atomc], dtype=np.int)
+restype_atomc_to_rigid_group = np.zeros([canonical_num_residues, num_atomc], dtype=np.int32)
 restype_atomc_mask = np.zeros([canonical_num_residues, num_atomc], dtype=np.float32)
 restype_atomc_rigid_group_positions = np.zeros([canonical_num_residues, num_atomc, 3], dtype=np.float32)
 restype_atom3_rigid_group_positions = np.zeros([canonical_num_residues, 3, 3], dtype=np.float32)
